+++ packages/web-ui/src/Book/UnitBookCount/styles.ts
@@ -1,6 +1,6 @@
-import { DisplayProperty, PositionProperty, WhiteSpaceProperty } from 'csstype';
+import { PositionProperty, WhiteSpaceProperty } from 'csstype';
 import { merge } from 'lodash';
-import { resetAppearance, resetFont, resetLayout } from "../../styles";
+import { resetAppearance, resetFont, resetLayout } from "src/styles";
 
 export const unitBookCount = merge({}, resetAppearance, resetFont, resetLayout, {
   display: 'inline-block',
@@ -47,7 +47,7 @@ export const num = {
 };
 
 export const icon = {
-  position: 'absolute',
+  position: 'absolute' as PositionProperty,
   right: 7,
   top: '50%',
   transform: 'translate3d(0, -50%, 0)',