+++ packages/web-ui/src/Book/UpdateBadge/index.tsx
@@ -1,8 +1,8 @@
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import { up as UpIcon} from '@ridi/web-icons';
 import * as React from 'react';
-import { displayHidden } from '../../styles';
+import { displayHidden } from 'src/styles';
+import UpIcon from 'src/svgs/Up.svg';
 import * as styles from './styles';
 
 export const UpdateBadge: React.FunctionComponent = () => (