+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.6.1] - 2019-08-12
+### Changed
+- Update the adult badge to use path instead of text
+
+## [0.6.0] - 2019-07-23
+### Changed
+- **BREAKING**: Book component only icons remove from '@ridi/web-icons'
+- Adult badge icon file type change : base64 → SVG
+  
+
 ## [0.5.0] - 2019-07-03
 ### Fixed
 - Fix max-height issue